requirements.txt
asgiref==3.2.10
autopep8==1.5.3
Django==3.0.8
django-timezone-field==4.0
pkg-resources==0.0.0
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
